nervus-cli/requirements.txt
textual>=0.47.0
httpx>=0.27.0
nats-py>=2.7.0
websockets>=12.0
pyaudio>=0.2.14
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.0.0
